Address adversarial native tool review
